[FLINK-19288] Make the InternalTimeServiceManager an interface
This refactoring allows to replace the InternalTimerService with a
different implementation.

Moreover it removes methods for retrieving the number of registered
timers from AbstractStreamOperator(s) which were introduced for tests.

This closes #13443
