+++ flink-streaming-java/src/main/java/org/apache/flink/streaming/api/operators/StreamTaskStateInitializerImpl.java
@@ -29,7 +29,6 @@ import org.apache.flink.runtime.checkpoint.PrioritizedOperatorSubtaskState;
 import org.apache.flink.runtime.checkpoint.StateObjectCollection;
 import org.apache.flink.runtime.execution.Environment;
 import org.apache.flink.runtime.jobgraph.OperatorID;
-import org.apache.flink.runtime.state.AbstractKeyedStateBackend;
 import org.apache.flink.runtime.state.CheckpointableKeyedStateBackend;
 import org.apache.flink.runtime.state.DefaultOperatorStateBackend;
 import org.apache.flink.runtime.state.KeyGroupRange;
@@ -89,23 +88,27 @@ public class StreamTaskStateInitializerImpl implements StreamTaskStateInitialize
 
 	private final TtlTimeProvider ttlTimeProvider;
 
+	private final InternalTimeServiceManager.Provider timeServiceManagerProvider;
+
 	public StreamTaskStateInitializerImpl(
 		Environment environment,
 		StateBackend stateBackend) {
 
-		this(environment, stateBackend, TtlTimeProvider.DEFAULT);
+		this(environment, stateBackend, TtlTimeProvider.DEFAULT, InternalTimeServiceManagerImpl::create);
 	}
 
 	@VisibleForTesting
 	public StreamTaskStateInitializerImpl(
 		Environment environment,
 		StateBackend stateBackend,
-		TtlTimeProvider ttlTimeProvider) {
+		TtlTimeProvider ttlTimeProvider,
+		InternalTimeServiceManager.Provider timeServiceManagerProvider) {
 
 		this.environment = environment;
 		this.taskStateManager = Preconditions.checkNotNull(environment.getTaskStateManager());
 		this.stateBackend = Preconditions.checkNotNull(stateBackend);
 		this.ttlTimeProvider = ttlTimeProvider;
+		this.timeServiceManagerProvider = timeServiceManagerProvider;
 	}
 
 	// -----------------------------------------------------------------------------------------------------------------
@@ -165,7 +168,12 @@ public class StreamTaskStateInitializerImpl implements StreamTaskStateInitialize
 			streamTaskCloseableRegistry.registerCloseable(rawOperatorStateInputs);
 
 			// -------------- Internal Timer Service Manager --------------
-			timeServiceManager = internalTimeServiceManager(keyedStatedBackend, keyContext, processingTimeService, rawKeyedStateInputs);
+			timeServiceManager = timeServiceManagerProvider.create(
+				keyedStatedBackend,
+				environment.getUserCodeClassLoader().asClassLoader(),
+				keyContext,
+				processingTimeService,
+				rawKeyedStateInputs);
 
 			// -------------- Preparing return value --------------
 
@@ -204,42 +212,6 @@ public class StreamTaskStateInitializerImpl implements StreamTaskStateInitialize
 
 			throw new Exception("Exception while creating StreamOperatorStateContext.", ex);
 		}
-	}
-
-	protected <K> InternalTimeServiceManager<K> internalTimeServiceManager(
-		CheckpointableKeyedStateBackend<K> keyedStatedBackend,
-		KeyContext keyContext, //the operator
-		ProcessingTimeService processingTimeService,
-		Iterable<KeyGroupStatePartitionStreamProvider> rawKeyedStates) throws Exception {
-
-		if (keyedStatedBackend == null) {
-			return null;
-		}
-
-		final KeyGroupRange keyGroupRange = keyedStatedBackend.getKeyGroupRange();
-		final boolean requiresSnapshotLegacyTimers = keyedStatedBackend instanceof AbstractKeyedStateBackend &&
-			((AbstractKeyedStateBackend<K>) keyedStatedBackend).requiresLegacySynchronousTimerSnapshots();
-
-		final InternalTimeServiceManager<K> timeServiceManager = new InternalTimeServiceManager<>(
-			keyGroupRange,
-			keyContext,
-			keyedStatedBackend,
-			processingTimeService,
-			requiresSnapshotLegacyTimers);
-
-		// and then initialize the timer services
-		for (KeyGroupStatePartitionStreamProvider streamProvider : rawKeyedStates) {
-			int keyGroupIdx = streamProvider.getKeyGroupId();
-
-			Preconditions.checkArgument(keyGroupRange.contains(keyGroupIdx),
-				"Key Group " + keyGroupIdx + " does not belong to the local range.");
-
-			timeServiceManager.restoreStateForKeyGroup(
-				streamProvider.getStream(),
-				keyGroupIdx, environment.getUserCodeClassLoader().asClassLoader());
-		}
-
-		return timeServiceManager;
 	}
 
 	protected OperatorStateBackend operatorStateBackend(
