+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/api/operators/StateInitializationContextImplTest.java
@@ -53,6 +53,7 @@ import org.apache.flink.runtime.state.TaskStateManagerImpl;
 import org.apache.flink.runtime.state.TestTaskLocalStateStore;
 import org.apache.flink.runtime.state.memory.ByteStreamStateHandle;
 import org.apache.flink.runtime.state.memory.MemoryStateBackend;
+import org.apache.flink.runtime.state.ttl.TtlTimeProvider;
 import org.apache.flink.runtime.taskmanager.CheckpointResponder;
 import org.apache.flink.runtime.util.LongArrayList;
 import org.apache.flink.streaming.runtime.tasks.ProcessingTimeService;
@@ -171,21 +172,22 @@ public class StateInitializationContextImplTest {
 		StateBackend stateBackend = new MemoryStateBackend(1024);
 		StreamTaskStateInitializer streamTaskStateManager = new StreamTaskStateInitializerImpl(
 			environment,
-			stateBackend) {
-
-			@Override
-			protected <K> InternalTimeServiceManager<K> internalTimeServiceManager(
-				CheckpointableKeyedStateBackend<K> keyedStatedBackend,
-				KeyContext keyContext,
-				ProcessingTimeService processingTimeService,
-				Iterable<KeyGroupStatePartitionStreamProvider> rawKeyedStates) throws Exception {
-
-				// We do not initialize a timer service manager here, because it would already consume the raw keyed
-				// state as part of initialization. For the purpose of this test, we want an unconsumed raw keyed
-				// stream.
-				return null;
-			}
-		};
+			stateBackend,
+			TtlTimeProvider.DEFAULT,
+			new InternalTimeServiceManager.Provider() {
+				@Override
+				public <K> InternalTimeServiceManager<K> create(
+						CheckpointableKeyedStateBackend<K> keyedStatedBackend,
+						ClassLoader userClassloader,
+						KeyContext keyContext,
+						ProcessingTimeService processingTimeService,
+						Iterable<KeyGroupStatePartitionStreamProvider> rawKeyedStates) throws Exception {
+					// We do not initialize a timer service manager here, because it would already consume the raw keyed
+					// state as part of initialization. For the purpose of this test, we want an unconsumed raw keyed
+					// stream.
+					return null;
+				}
+			});
 
 		AbstractStreamOperator<?> mockOperator = mock(AbstractStreamOperator.class);
 		when(mockOperator.getOperatorID()).thenReturn(operatorID);
