+++ flink-streaming-java/src/test/java/org/apache/flink/streaming/api/operators/StreamTaskStateInitializerImplTest.java
@@ -283,16 +283,19 @@ public class StreamTaskStateInitializerImplTest {
 		} else {
 			return new StreamTaskStateInitializerImpl(
 				dummyEnvironment,
-				stateBackend) {
-				@Override
-				protected <K> InternalTimeServiceManager<K> internalTimeServiceManager(
-					CheckpointableKeyedStateBackend<K> keyedStatedBackend,
-					KeyContext keyContext,
-					ProcessingTimeService processingTimeService,
-					Iterable<KeyGroupStatePartitionStreamProvider> rawKeyedStates) throws Exception {
-					return null;
-				}
-			};
+				stateBackend,
+				TtlTimeProvider.DEFAULT,
+				new InternalTimeServiceManager.Provider() {
+					@Override
+					public <K> InternalTimeServiceManager<K> create(
+							CheckpointableKeyedStateBackend<K> keyedStatedBackend,
+							ClassLoader userClassloader,
+							KeyContext keyContext,
+							ProcessingTimeService processingTimeService,
+							Iterable<KeyGroupStatePartitionStreamProvider> rawKeyedStates) throws Exception {
+						return null;
+					}
+				});
 		}
 	}
 }
